Text.EDE.Filters

Description

The means to construct your own filters.

Prelude

The default filters available to a template is documented by the following categories.

These filters cannot be overriden and attempting to supply your own filters to renderWith will cause the similarly named filters to disappear when they are merged with the prelude during evaluation. (See: union)

Boolean

! :: Bool -> Bool (See: not)

&& :: Bool -> Bool -> Bool

|| :: Bool -> Bool -> Bool

Equality

== :: a -> a -> Bool

!= :: a -> a -> Bool (See: /=)

Relational

> :: a -> a -> Bool

>= :: a -> a -> Bool

<= :: a -> a -> Bool

<= :: a -> a -> Bool

Numeric

+ :: Number -> Number -> Number

- :: Number -> Number -> Number

* :: Number -> Number -> Number

abs :: Number -> Number

signum :: Number -> Number

negate :: Number -> Number

Fractional

truncate :: Number -> Number

round :: Number -> Number

ceiling :: Number -> Number

floor :: Number -> Number

Textual

takeWord :: Text -> Text

dropWord :: Text -> Text

lowerHead :: Text -> Text

upperHead :: Text -> Text

toTitle :: Text -> Text

toCamel :: Text -> Text

toPascal :: Text -> Text

toSnake :: Text -> Text

toSpinal :: Text -> Text

toTrain :: Text -> Text

toLower :: Text -> Text

toUpper :: Text -> Text

toOrdinal :: Number -> Text

Collection

length :: Collection -> Number (See: length, length, size)

empty :: Collection -> Bool (See: null, null, null)

Polymorphic

show :: a -> Text

Constructing filters

data Binding

A HOAS representation of (possibly partially applied) values in the environment.

Constructors

BVal !Value

BLam (Binding -> Result Binding)

Classes

class Quote a where

Methods

quote :: a -> Binding

class Unquote a where

Methods

unquote :: Binding -> Result a

Restricted quoters

qapply :: Binding -> Binding -> Result Binding

Attempt to apply two Bindings.

qpoly2 :: Quote a => (Value -> Value -> a) -> Binding

Quote a binary function which takes the most general binding value.

qnum1 :: (Scientific -> Scientific) -> Binding

Quote an unary numeric function.

qnum2 :: Quote a => (Scientific -> Scientific -> a) -> Binding

Quote a binary numeric function.

qcol1 :: Quote a => (Text -> a) -> (Object -> a) -> (Array -> a) -> Binding

Quote a comprehensive set of unary functions to create a binding that supports all collection types.

Errors

unexpected :: String -> String -> Result b

typeOf :: Value -> String

Retrieve a consistent type from a Value to use in error messages.