Text.EDE

Description

A (mostly logicless) textual templating language with similar syntax to Liquid or Jinja2.

(ED-E is a character from Fallout New Vegas, pronounced iː-diː-iː, or Eddie.)

How to use this library

A simple example of parsing and rendering Text containing a basic conditional expression and variable interpolation follows.

First the Template is defined:

>>> let tmpl = "{% if var %}\nHello, {{ var }}!\n{% else %}\nnegative!\n{% endif %}\n" :: Data.ByteString.ByteString

Then an Object is defined containing the environment which will be available to the Template during rendering:

>>> let env = fromPairs [ "var" .= "World" ] :: Object

Note: the fromPairs function above is a wrapper over Aeson's object which removes the Value constructor, exposing the delicious HashMap underneath.

Finally the environment is applied to the Template:

>>> render tmpl env :: Result Text
> Success "Hello, World!"

In this manner, Templates can be pre-compiled to the internal AST and the cost of parsing can be amortised if the same Template is rendered multiple times.

Another example, this time rendering a Template from a file:

import qualified Data.Text.Lazy as LText
import           Text.EDE

main :: IO ()
main = do
    r <- eitherParseFile "template.ede"
    either error print $ r >>= (`eitherRender` env)
  where
    env = fromPairs
        [ "text" .= "Some Text."
        , "int"  .= 1
        , "list" .= [5..10]
        ]

Please see the syntax section for more information about available statements and expressions.

Parsing and Rendering

Parsing and rendering require two separate steps intentionally so that the more expensive (and potentially impure) action of parsing and resolving includes can be embedded and re-used in a pure fashion.

Parsing tokenises the input and converts it to an internal AST representation, resolving includes using a custom function. The result is a compiled template which can be cached for future use.

Rendering takes a HashMap of custom Funs (functions available in the template context), an Object as the binding environment, and a parsed Template to subsitute the values into. The result is a Lazy Text value containing the rendered output.

data Template

A parsed and compiled template.

Parsing

parse

Arguments

:: ByteString
Strict ByteString template definition.

-> Result Template

Parse Lazy Text into a compiled Template.

Because this function is pure and does not resolve includes, encountering an include expression during parsing will result in an Error.

See parseFile or parseWith for mechanisms to deal with include dependencies.

parseIO

Arguments

:: FilePath
Parent directory for relatively pathed includes.

-> ByteString
Strict ByteString template definition.

-> IO (Result Template)

Parse Text into a compiled Template.

This function handles all include expressions as FilePaths and performs recursive loading/parsing.

parseFile

Arguments

:: FilePath
Path to the template to load and parse.

-> IO (Result Template)

Load and parse a Template from a file.

This function handles all include expressions as FilePaths and performs recursive loading/parsing, with pathing of includes relatively to the target (unless absolute paths are used).

parseFileWith

Arguments

:: Syntax
Delimiters and parsing options.

-> FilePath
Path to the template to load and parse.

-> IO (Result Template)

See: parseFile.

parseWith

Arguments

:: Monad m

=> Syntax
Delimiters and parsing options.

-> Resolver m
Function to resolve includes.

-> Text
Strict Text name.

-> ByteString
Strict ByteString template definition.

-> m (Result Template)

Parse a Template from a Strict ByteString using a custom function for resolving include expressions.

Two custom include resolvers are supplied:

includeMap

includeFile

parseFile for example, is defined as: parseWith includeFile.

Includes

The Resolver used to resolve include expressions determines the purity of Template parsing.

For example, using the includeFile Resolver means parsing is restricted to IO, while pre-caching a HashMap of Templates and supplying them to parseWith using includeMap offers a pure variant for include resolution.

type Resolver m = Syntax -> Text -> Delta -> m (Result Template)

A function to resolve the target of an include expression.

includeMap

Arguments

:: Monad m

=> HashMap Text Template
A HashMap of named Templates.

-> Resolver m
Resolver for parseWith.

HashMap resolver for include expressions.

The identifier component of the include expression is treated as a lookup key into the supplied HashMap. If the identifier doesn't exist in the HashMap, an Error is returned.

includeFile

Arguments

:: FilePath
Parent directory for relatively pathed includes.

-> Resolver IO

FilePath resolver for include expressions.

The identifier component of the include expression is treated as a relative FilePath and the template is loaded and parsed using parseFile. If the identifier doesn't exist as a valid FilePath, an Error is returned.

Filters

defaultFilters :: HashMap Text Binding

Rendering

render

Arguments

:: Template
Parsed Template to render.

-> Object
Bindings to make available in the environment.

-> Result Text

Render an Object using the supplied Template.

renderWith

Arguments

:: HashMap Text Binding
Filters to make available in the environment.

-> Template
Parsed Template to render.

-> Object
Bindings to make available in the environment.

-> Result Text

Render an Object using the supplied Template.

Either Variants

eitherParse :: ByteString -> Either String Template

See: parse

eitherParseFile :: FilePath -> IO (Either String Template)

See: parseFile

eitherParseWith :: (Functor m, Monad m) => Syntax -> Resolver m -> Text -> ByteString -> m (Either String Template)

See: parseWith

eitherRender :: Template -> Object -> Either String Text

See: render

eitherRenderWith :: HashMap Text Binding -> Template -> Object -> Either String Text

See: renderWith

Results and Errors

The Result of a parse or render steps can be inspected or analysed using result as follows:

>>> result failure success $ render tmpl env

If you're only interested in dealing with errors as strings, and the positional information contained in Meta is not of use you can use the convenience functions eitherParse, eitherRender, or convert a Result to Either using eitherResult.

>>> either failure success $ eitherParse tmpl

data Delta :: *

Constructors

Columns !Int64 !Int64

Tab !Int64 !Int64 !Int64

Lines !Int64 !Int64 !Int64 !Int64

Directed !ByteString !Int64 !Int64 !Int64 !Int64

data Result a

The result of running parsing or rendering steps.

Constructors

Success a

Failure Doc

eitherResult :: Result a -> Either String a

Convert a Result to an Either with the Left case holding a formatted error message, and Right being the successful result over which Result is paramterised.

result

Arguments

:: (Doc -> b)
Function to apply to the Failure case.

-> (a -> b)
Function to apply to the Success case.

-> Result a
The Result to map over.

-> b

Perform a case analysis on a Result.

success :: Monad m => a -> m (Result a)

Convenience for returning a successful Result.

failure :: Monad m => Doc -> m (Result a)

Convenience for returning an error Result.

Input

fromPairs (or fromValue) is a wrapper around Aeson's object function which safely strips the outer Value constructor, providing the correct type signature for input into render.

It is used in combination with the re-exported .= as follows:

>>> render (fromPairs [ "foo" .= "value", "bar" .= 1 ]) :: Template -> Result Text

fromValue :: Value -> Maybe Object

Unwrap a Value to an Object safely.

See Aeson's documentation for more details.

fromPairs :: [Pair] -> Object

Create an Object from a list of name/value Pairs.

See Aeson's documentation for more details.

(.=) :: ToJSON a => Text -> a -> Pair

Version

version :: Version

ED-E Version.

Syntax

type Delim = (String, String)

data Syntax

delimRender :: Lens' Syntax Delim

delimComment :: Lens' Syntax Delim

delimBlock :: Lens' Syntax Delim

defaultSyntax :: Syntax

The default ED-E syntax.

Delimieters:

Inline: {{ ... }}

Comments: {# ... #}

Control Structures: {% ... %}

alternateSyntax :: Syntax

An alternate syntax (based on Play/Scala templates) which is designed to be used when the default is potentially ambiguous due to another smarty based template syntax.

Delimiters:

Inline: <@ ... @>

Comments: @* ... *@

Control Structures: @( ... )@

Comments

Comments are ignored by the parser and omitted from the rendered output.

{# singleline comment #}

{#
   multiline
   comment
#}

Raw

You can disable template processing for blocks of text using the raw section:

{% raw %}
Some {{{ handlebars }}} or {{ mustache }} or {{ jinja2 }} output tags etc.
{% endraw %}

This can be used to avoid parsing expressions which would otherwise be considered valid ED-E syntax.

Variables

Variables are substituted directly for their Buildable representation. An error is raised if the varaible being substituted is not a literal type (ie. an Array or Object) or doesn't exist in the supplied environment.

{{ var }}

Nested variable access is also supported for variables which resolve to an Object. Dot delimiters are used to chain access through multiple nested Objects. The right-most accessor must resolve to a Buildable type as with the previous non-nested variable access.

{{ nested.var.access }}

Conditionals

A conditional is introduced and completed with the section syntax:

{% if var %}
   ... consequent expressions
{% else %}
   ... alternate expressions
{% endif %}

The value of {{ var }} determines the branch that is rendered by the template with the else branch being optional.

In the case of a literal it conforms directly to the supported boolean or relation logical operators from Haskell. If a variable is singuarly used its existence determines the result of the predicate, the exception to this rule is boolean values which will be substituted into the expression if they exist in the supplied environment.

The following logical expressions are supported as predicates in conditional statements with parameters type checked and an error raised if the left and right hand sides are not type equivalent.

And: &&

Or: ||

Equal: ==

Not Equal: /=

Greater: >

Greater Or Equal: >=

Less: <

Less Or Equal: <=

Negation: !

Case Analysis

To pattern match a literal or variable, you can use the case statement:

{% case var %}
{% when "a" %}
   .. matched expressions
{% when "b" %}
   .. matched expressions
{% else %}
   .. alternate expressions
{% endcase %}

Loops

Iterating over an Array or Object can be acheived using the 'for ... in' section syntax. Attempting to iterate over any other type will raise an error.

Example:

{% for var in list %}
    ... iteration expression
{% else %}
    ... alternate expression
{% endfor %}

The iteration branch is rendering per item with the else branch being (which is optional) being rendered if the {{ list }} variable is empty.

When iterating over an Object, a stable sort using key equivalence is applied, Arrays are unmodified.

The resulting binding within the iteration expression (in this case, {{ var }}) is an Object containing the following keys:

key :: Text: They key if the loop target is an Object

value :: a: The value of the loop target

loop :: Object: Loop metadata.

length :: Int: Length of the loop

index :: Int: Index of the iteration

index0 :: Int: Zero based index of the iteration

remainder :: Int: Remaining number of iterations

remainder0 :: Int: Zero based remaining number of iterations

first :: Bool: Is this the first iteration?

last :: Bool: Is this the last iteration?

odd :: Bool: Is this an odd iteration?

even :: Bool: Is this an even iteration?

For example:

{% for item in items %}
    {{ item.index }}:{{ item.value }}
    {% if !item.last %}

    {% endif %}
{% endfor %}

Will render each item with its (1-based) loop index as a prefix, separated by a blank newline, without a trailing at the end of the document.

Includes

Includes are a way to reduce the amount of noise in large templates. They can be used to abstract out common snippets and idioms into partials.

If parseFile or the includeFile resolver is used, templates will be loaded from FilePaths, for example:

{% include "/var/tmp/partial.ede" %}

Loads partial.ede from the file system.

By default, the current environment is made available to the included template, but this can be overriden by specifying a specific binding to make available:

{% include "/var/tmp/partial.ede" with value %}

Will ensure only the key value (and descendents) is available in the partial's environment.

Includes can also be resolved using pure Resolvers such as includeMap, which will treat the include expression's identifier as a HashMap key:

{% include "arbitrary_key" %}

Uses lookup to find arbitrary_key in the HashMap supplied to includeMap.

Filters

Filters are typed functions which can be applied to variables and literals. An example of rendering a lower cased boolean would be:

{{ true | show | lower }}

The input is on the LHS and chained filters (delimited by '|') are on the RHS, with filters being applied left associatively.

See: Text.EDE.Filters.