Text.EDE.Filters

Description

A default set of prelude-like filters and the means to construct and add your own.

Defaults

defaultFilters :: HashMap Text Binding

Boolean

!  :: Bool -> Bool

&& :: Bool -> Bool -> Bool

|| :: Bool -> Bool -> Bool

Equality

== :: a -> a -> Bool

!= :: a -> a -> Bool

See: '(==)'

Relation

> :: a -> a -> Bool

>= :: a -> a -> Bool

:: a - a -> Bool

:: a - a -> Bool

Numeric

+      :: Number -> Number -> Number

-      :: Number -> Number -> Number

*      :: Number -> Number -> Number

abs    :: Number -> Number

signum :: Number -> Number

negate :: Number -> Number

Fractional

truncate :: Number -> Number

round    :: Number -> Number

ceiling  :: Number -> Number

floor    :: Number -> Number

Textual

lower      :: Text -> Text

upper      :: Text -> Text

lowerFirst :: Text -> Text

upperFirst :: Text -> Text

titleize   :: Text -> Text

pascalize  :: Text -> Text

camelize   :: Text -> Text

underscore :: Text -> Text

hyphenate  :: Text -> Text

Collection

length :: Collection a => a -> Number

empty :: Collection a => a -> Bool

Polymorphic

show :: a -> Text

Constructing filters

data Binding

A HOAS representation of (possibly partially applied) values in the environment.

Constructors

BVal !Value

BLam (Binding -> Result Binding)

Classes

class Quote a where

Methods

quote :: a -> Binding

class Unquote a where

Methods

unquote :: Binding -> Result a

Restricted quoters

qapply :: Binding -> Binding -> Result Binding

Attempt to apply two Bindings.

qpoly2 :: Quote a => (Value -> Value -> a) -> Binding

Quote a binary function which takes the most general binding value.

qnum1 :: (Scientific -> Scientific) -> Binding

Quote an unary numeric function.

qnum2 :: Quote a => (Scientific -> Scientific -> a) -> Binding

Quote a binary numeric function.

qcol1 :: Quote a => (Text -> a) -> (Object -> a) -> (Array -> a) -> Binding

Quote a comprehensive set of unary functions to create a binding that supports all collection types.

Errors

unexpected :: String -> String -> Result b

typeOf :: Value -> String

Retrieve a consistent type from a Value to use in error messages.